quicSocket to its own file
