+++ transport/socket.go
@@ -14,7 +14,6 @@ import (
 	"strings"
 	"time"
 
-	"github.com/lucas-clemente/quic-go"
 	"github.com/ortuman/jackal/transport/compress"
 )
 
@@ -112,31 +111,4 @@ func (s *socketTransport) PeerCertificates() []*x509.Certificate {
 		return st.PeerCertificates
 	}
 	return nil
-}
-
-type quicSocketTransport struct {
-	socketTransport
-	conn quic.Session
-}
-
-func NewQUICSocketTransport(conn quic.Session, uniStream quic.Stream,
-	keepAlive time.Duration) Transport {
-	s := &quicSocketTransport{
-		socketTransport: socketTransport{
-			rw:        uniStream,
-			br:        bufio.NewReaderSize(uniStream, socketBuffSize),
-			bw:        bufio.NewWriterSize(uniStream, socketBuffSize),
-			keepAlive: keepAlive,
-		},
-		conn: conn,
-	}
-	return s
-}
-
-func (s *quicSocketTransport) Read(p []byte) (n int, err error) {
-	n, err = s.br.Read(p)
-	return n, err
-}
-
-func (s *quicSocketTransport) StartTLS(cfg *tls.Config, asClient bool) {
 }
